wip: Move rundown timing to lib and add tests
